Track time since last DE search; extract metrics to finer module
